修改了maven依赖，使得能够正常运行 完善了线程级别用户信息存储

+++ meowcloud-common/src/main/java/com/xhpolaris/meowcloud/common/model/vo/MemoryVO.java
@@ -2,20 +2,19 @@ package com.xhpolaris.meowcloud.common.model.vo;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import org.checkerframework.checker.units.qual.A;
 
 @ApiModel("存储信息展示")
 public class MemoryVO {
 
-    // TODO 已使用内存的展示模式，初步计划为1K为最小存储单位
+    // TODO 已使用内存的展示模式，初步计划为1Byte为最小存储单位
     @ApiModelProperty("已使用内存")
-    private Integer usedMemory;
+    private Long usedMemory;
 
     @ApiModelProperty("相片总数")
-    private Integer photoTotal;
+    private Long photoTotal;
 
     @ApiModelProperty("相册总数")
-    private Integer albumTotal;
+    private Long albumTotal;
 
     // 以1%为最小单位
     @ApiModelProperty("已使用的比例")
